src/concrete-core/pyproject.toml
[project]
name = "concrete-core"
version = "0.0.7"
description = "AI Agent Orchestration"
readme = "README.md"
requires-python = ">=3.11.9, <4.0.0"

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"] # https://github.com/openai/openai-python/releases/tag/v1.40.0
requests = ["requests>=2.0.0"]  
auth = [
    "pyjwt>=2.9.0",
    "cryptography>=35.0.0",
    "itsdangerous>=2.0.0",  # session middleware
    "fastapi>=0.105"
]
google-auth = [
    "google-auth>=2.25",
    "google-auth-oauthlib>=1.1.1",
    "google-api-python-client>=2.0.0"
]
otel=[ 
    "opentelemetry-api>=1.28.2",
    "opentelemetry-exporter-otlp>=1.28.2",
    "opentelemetry-propagator-b3>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["concrete"]
